refactor: 💡 Rename files and functions to better reflect what
they do. Also rename `data-speccer*`-attributes to make the API a bit
easier to handle and understand. This commit also comes with some more
guarding for the feature functions, and added some more tests.

BREAKING CHANGE: 🧨 The API has changed! Please refer to the `README.md` and api
documentation for this! You should migrate easily to the current version
by search and replace.

✅ Closes: #225