refactor: 💡 Use options object instead of area strings
To be able to steer the API towards a decent interface for programmatic
usage and other frameworks, the settings and features enabled are to be
governed by passed options instead of strings.

BREAKING CHANGE: 🧨 Have changed the API to accept an option object instead of a string.
Please check the documentation.